// src/api/generated/models/authControllerLogin200.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * devit-api
 * Devit API
 * OpenAPI spec version: 0.0.1
 */
import type { AuthControllerLogin200User } from './authControllerLogin200User';

export type AuthControllerLogin200 = {
  token?: string;
  user?: AuthControllerLogin200User;
};